LongToInteger

+++ back/src/main/java/fr/repository/RegistrationRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import fr.entity.Registration;
 
-public interface RegistrationRepository extends JpaRepository<Registration, Long> {
+public interface RegistrationRepository extends JpaRepository<Registration, Integer> {
     
     @Query("SELECT r FROM Registration r WHERE r.prestation.id = :id")
-    Iterable<Registration> findAllByPrestationId(@Param("id") Long id);
+    Iterable<Registration> findAllByPrestationId(@Param("id") Integer id);
 
     @Query("SELECT r, p FROM Registration r JOIN Prestation p")
     Iterable<Registration> findAllPrestationWithComment();
